requirements.txt
amqp==5.1.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.3.0
Flask==2.3.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
numpy==1.25.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==23.1
pip-autoremove==0.10.0
pluggy==1.2.0
prompt-toolkit==3.0.39
pymongo==4.4.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.6.0
six==1.16.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
